用户模块新增删除、修改接口开发

+++ src/main/java/com/yi/easycode/modules/user/mapper/UserMapper.java
@@ -16,5 +16,5 @@ public interface UserMapper extends BaseMapper<UserEntity> {
 
     UserEntity getUserInfoByName(String name);
 
-    UserVO getUserInfo(Integer userId);
+    UserVO getUserInfo(Long userId);
 }
